UEFI-pad werkend op de Ampere Altra: bootloader, ACPI, igb, TRNG
Het universele serverpad (ontwerpprincipe: universeel boven
board-specifiek) — één BOOTAA64.EFI voor elk AArch64-platform met
UEFI+ACPI, de Altra (128 cores) voorop, QEMU+EDK2 als proeftuin:

- UEFI-bootloader: mkkernel -pe wikkelt de kernel als PE/COFF;
  init.s draait als UEFI-app (ImageHandle/SystemTable), haalt de
  memory map op, ExitBootServices, en neemt de machine over;
  board/uefi met eigen 48-bit MMU-plan en UEFI-watchdog-afmelding
- metal/acpi: RSDP→XSDT-parser — MADT (cores), MCFG (PCIe-ECAM),
  SPCR (console-UART), FADT (PSCI-conduit); het ACPI-equivalent
  van metal/fdt, alleen lezen, geen AML
- metal/igb: driver voor de Intel igb-familie — de I210 op de
  Altra Dev Kit, registerlaag end-to-end bewezen tegen QEMU's
  82576-model; polled, advanced descriptors, zelfde
  NetworkDevice-vorm als gem/genet/virtionet
- metal/trng: hardware-entropie via FEAT_RNG (RNDR) of SMCCC TRNG
  (TF-A op de Altra), met aflopende voorkeur
- docs/uefi.md + docs/app-memory.md publiek; tools/test.sh als
  host-test + tamago-compile-gate (virt/rpi4/rpi5/uefi)

+++ metal/stage2/stage2_test.go
@@ -252,7 +252,7 @@ func TestIsolatieTussenSlots(t *testing.T) {
 		{"ctrl-page", uint64(layout.CtrlPagePA(2)), layout.CtrlStride},
 		{"hop-ringen", uint64(layout.RingOutboxPA(2)), layout.RingStride},
 		{"net-ringen", uint64(layout.NetRingTXPA(2)), layout.NetRingStride},
-		{"stage-2-tabellen", uint64(layout.Stage2TablePA(0)), (layout.MaxSlots + 1) * layout.Stage2Stride},
+		{"stage-2-tabellen", uint64(layout.Stage2TablePA(0)), uint64(layout.MaxSlots+1) * layout.Stage2Stride},
 	}
 	roGezien := 0
 	for _, l := range walk(t, 1) {
